seek streamed game file + notice

+++ CUE4Parse/UE4/IO/Objects/FIoStoreTocResource.cs
@@ -40,7 +40,7 @@ namespace CUE4Parse.UE4.IO.Objects
             if (Header.Version < EIoStoreTocVersion.PartitionSize)
             {
                 Header.PartitionCount = 1;
-                Header.PartitionSize = uint.MaxValue;
+                Header.PartitionSize = ulong.MaxValue;
             }
 
             // Chunk IDs
